Change add link to row approach
This approach allows the user to take full advantage of the link that the <a> tag provides, including opening in a new window with a middle click or the expected items in the context menu. For example, a middle click, which is programmatically configured using js, moved the user to a new tab instead of leaving him on the tab where the click was made.